+++ p3backend/ReportsService/src/main/java/com/revature/ReportsService/service/DataGeneratorService.java
@@ -2,9 +2,7 @@ package com.revature.ReportsService.service;
 
 import java.util.ArrayList;
 import java.util.List;
-
 import org.springframework.stereotype.Service;
-
 import com.revature.ReportsService.models.Associate;
 
 @Service
@@ -22,7 +20,7 @@ public class DataGeneratorService {
 
   private List<Associate> associatesInOneBatch;
 
-  public List<Associate> generateBatch(int interviewScore, int quantity, List<Associate> aList) {
+  public List<Associate> generateBatch(Double interviewScore, int quantity, List<Associate> aList) {
     associatesInOneBatch = new ArrayList<Associate>();
     // associatesInOneBatch = null;
